+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - started working on the lexer
 
 ## 2022/05/26
--
+- finished a first version of the lexer
+- finished a stub of the parser